refactor: move custom exceptions from constants.py to exceptions.py
