+++ mip_template/constants.py
@@ -14,18 +14,3 @@ class _SampleConstants(NamedTuple):
 
 # e.g. SAMPLE_CONSTANTS.FIRST_VALUE returns the string 'Value 1'. It can be used elsewhere to check equality.
 SAMPLE_CONSTANTS = _SampleConstants()
-
-
-class InputDataError(Exception):
-    """
-    Raised whenever the input data is not valid.
-    """
-
-
-class SolutionError(Exception):
-    """
-    Raised when the output from the optimization or the output tables contain some problem.
-    
-    For example, if the solution is not feasible, or if the output from the optimization violates an expected
-    constraint.
-    """
